Refactor business details route
https://eaflood.atlassian.net/browse/SFDP-554

This PR refactors the first of the business details pages. Following team discussions, we've agreed to keep the route definitions (GET and POST) and their corresponding handlers together, while moving the business logic within the handlers into private service modules.

The goal is to reduce complexity in the routes, improve code reusability, and ensure unit testing is robust and straightforward.