VsCode Tasks Config (#51)
This PR is a chore for adding VSCode tasks for common frontend Docker commands.

Adds a set of developer tasks to tasks.json for quickly managing the build of the Docker container. These tasks include up, down, log, test, lint and clean. These tasks streamline local dev workflows and reduce the need to manually run the commands on the terminal.