Fix broken back links #patch (#70)
This PR fixes broken backlinks across the service that haven't yet been updated as part of the refactoring work. For now, most backlinks will direct users to the Business Details page. Once additional pages are built and the service's navigation is updated, these links can be revised to take users back to the appropriate previous pages.