Remove validation on DAL data after fetch (#110)
During testing of the DAL connector, it was found that many test logins fail to load the business details or personal details pages. This happens because the data fetched from the DAL is being validated against our schema, and many fields fail this validation. As a result, users are taken to the “There is a problem with this service” page.

Because of this, users cannot access pages to update incorrect data via the “change” link. This PR removes the initial validation that runs immediately after fetching DAL data.

Additionally, this PR adds a sign-out link to the home page. On the test environment, if a user’s personal and business details pages fail to load, they previously had no way to log out and try a different account