Update business phone numbers to use dalConnector
https://eaflood.atlassian.net/browse/SFD2-37

This PR adds in the functionality for the updateBusinessPhoneNumberService to connect to the DAL to save landline and mobile numbers. The key changes are:

Integration with the DAL connector in updateBusinessPhoneNumberService. This takes in the mutation query and the variables needed to send to the DAL.

Displays a success banner after a successful update.

Throws an error if the DAL fails or returns an unsuccessful status.