business VAT template filename case for test environment (#137)
https://eaflood.atlassian.net/browse/SFD2-277

Fix: business VAT template filename case for test environment

PR Description:
This PR fixes a 500 error on the /business-vat-registration-number-change, /business-vat-registration-number-remove and /business-vat-registration-number-check pages in the test environment.

Issue:
The template file was named business-VAT-registration-number-change.njk (uppercase “VAT”), while the route referenced business-vat-registration-number-change. This worked locally (case-insensitive filesystem) but caused a 500 error on the test environments (case-sensitive filesystem).

Fix:
Renamed the template file to business-vat-registration-number-change.njk to match the route.

We have verified the page now renders correctly in the test environment using the PR as a hotfix.