Change business address manual enter presenter (#159)
https://eaflood.atlassian.net/browse/SFD2-40

After testing the business address pages, several bugs and changes in behaviour were identified.

Changes made:

Error handling on inputs: Previously, validation errors were not displayed on the view. This has been fixed by adding the appropriate error properties to the GOV.UK input components.
Country capitalisation: Country values from the address lookup were being returned in lowercase, while the rest of the address was capitalised. This has now been updated so all address fields are consistently formatted.
Lookup addresses mapped to manual fields: Following discussions with the design and testing teams, it was decided that when a user chooses to manually enter an address, any existing lookup address should be mapped to the manual entry fields. This ensures the user sees the current address that is saved for them.