Update: align right macros (#175)
https://eaflood.atlassian.net/browse/SFD2-465

During a design review for another ticket, the design team noticed that the user name and “Sign out” link were not correctly aligned to the right side of the page, as shown in the prototype (see screenshot of current behaviour).

This PR adds the missing -!- prefix to the align-right class on both elements so they now align correctly on the page.