SFD2-502 Fix text wrapping for business and user name (#186)
https://eaflood.atlassian.net/browse/SFD2-502

This PR adds a custom CSS class to fix the text wrapping issue affecting user names and business names when they exceed a certain visual length.

There isn’t a fixed character limit for when wrapping occurs, as it varies depending on:

Whether the service is viewed on the web or mobile, and
The specific characters used (e.g. a capital “W” takes up more space than a lowercase “i”).
This change ensures that text wraps dynamically based on the container width, improving layout consistency across devices and screen sizes.

Additional housekeeping tasks done:

Removed redundant permissions code in the server.js file. This code was running a pre-auth function that console logged out when a user didn't have the correct permissions. This was done to help debug login requests failing before the full permissions work was finished. Now that work is done, this code can be removed (it was forgotten about in the permissions work PR, opps)