SFD2-363: Add route protection to personal details change routes
https://eaflood.atlassian.net/browse/SFD2-363

Currently all the business details pages have route protection based on user permissions. However, the personal details change pages do not have this protection in place. This means that users without the correct permissions can access these pages, which could lead to unauthorized changes to personal details.
This PR adds route protection to the personal details change pages. The user must now have the 'view' permission to access these routes.

The following pages in the service do not have protected routes:

* Footer contact us page
* Footer accessibility page
* Footer cookies page

This is because a user should be able to access these pages without needing to be authenticated or have specific permissions.