Fix footer routes auth
https://eaflood.atlassian.net/browse/SFD2-363

As part of the work to fix the personal details routes authentication this PR is tackling the footer routes.
Anybody should be able to access the footer routes (contact us, accessibility statement and cookies) without being authenticated.
This PR adds the necessary changes to the routes and tests to achieve this.